Move NodeAtPosition from Index to Utility
