+++ phpstan.neon
@@ -22,7 +22,7 @@ parameters:
             message: 'AST traversal is confined: parser pipeline (ParserService), positional finders (NodeAtPosition, Scope, ScopeFinder), and DeclarationScanner'
             allowIn:
                 - src/Parser/ParserService.php
-                - src/Index/NodeAtPosition.php
+                - src/Utility/NodeAtPosition.php
                 - src/Knowledge/DeclarationScanner.php
                 - src/Utility/Scope.php
                 - src/Utility/ScopeFinder.php
@@ -32,7 +32,7 @@ parameters:
             message: 'AST traversal is confined: parser pipeline (ParserService), positional finders (NodeAtPosition, Scope, ScopeFinder), and DeclarationScanner'
             allowIn:
                 - src/Parser/ParserService.php
-                - src/Index/NodeAtPosition.php
+                - src/Utility/NodeAtPosition.php
                 - src/Knowledge/DeclarationScanner.php
                 - src/Utility/Scope.php
                 - src/Utility/ScopeFinder.php
@@ -64,7 +64,7 @@ parameters:
             message: 'AST traversal is confined: parser pipeline (ParserService), positional finders (NodeAtPosition, Scope, ScopeFinder), and DeclarationScanner'
             allowIn:
                 - src/Parser/ParserService.php
-                - src/Index/NodeAtPosition.php
+                - src/Utility/NodeAtPosition.php
                 - src/Knowledge/DeclarationScanner.php
                 - src/Utility/Scope.php
                 - src/Utility/ScopeFinder.php
