+++ tests/Index/NodeAtPositionTest.php
@@ -5,7 +5,7 @@ declare(strict_types=1);
 namespace Firehed\PhpLsp\Tests\Index;
 
 use Firehed\PhpLsp\Document\TextDocument;
-use Firehed\PhpLsp\Index\NodeAtPosition;
+use Firehed\PhpLsp\Utility\NodeAtPosition;
 use Firehed\PhpLsp\Parser\ParserService;
 use PhpParser\Node;
 use PhpParser\Node\Identifier;
